+++ packages/num/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zinaid/num
 
+## 0.0.9
+
+### Patch Changes
+
+- Small fixes to package to modularize packages
+
+## 0.0.8
+
+### Patch Changes
+
+- Test release package
+- Updated dependencies
+  - @zinaid/utils@0.0.5
+
 ## 0.0.7
 
 ### Patch Changes