+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zinaid/types
 
+## 0.0.4
+
+### Patch Changes
+
+- Test release package
+
 ## 0.0.3
 
 ### Patch Changes